requirements.txt
# Core Data Processing
pandas>=1.5.0

# Google Sheets Integration
gspread>=5.0.0
oauth2client>=4.1.3

# HTTP Requests (for LIV25 API)
requests>=2.28.0

# Date/Time Parsing
python-dateutil>=2.8.0

# Optimization (Gurobi)
gurobipy>=10.0.0


